Refactor import statements in workflow.py and debug_test.py
