+++ tests/profile_workflow.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from amazon_scraper.configuration.inject import ConfigStore, ConfigValue
+from amazon_scraper.configuration import ConfigStore, ConfigValue
 from amazon_scraper.scripts.main import scrape_workflow
 
 ConfigStore.configure_context(source='tests/profile_config.yml')
